Start moving things about
Put each implementation in its own file, sort of.
