+++ Makefile
@@ -1,12 +1,46 @@
+OS_NAME = $(shell uname)
+MACH_NAME = $(shell uname -m)
+
+ifeq ($(OS_NAME), Darwin)
+OS = OSX
+endif
+
+ifeq ($(OS_NAME), Linux)
+OS = LINUX
+endif
+
+# Mac OSX
+ifeq ($(OS), OSX)
+CFLAGS = -I/opt/local/include
+LDFLAGS = -L/opt/local/include
+endif
+
 all: cbridge
 
-# YMMV, but sometimes openssl etc are in /opt/local. -lssl and -lcrypto are needed only for TLS
-cbridge: cbridge.c chaosd.h cbridge-chaos.h chudp.h
-	$(CC) -I/opt/local/include -L/opt/local/lib $(CFLAGS) -o cbridge cbridge.c -lpthread -lssl -lcrypto
+OBJS = contacts.o usockets.o chtls.o chudp.o debug.o chether.o
 
-# Just a PoC
-uttun: uttun.c
-	$(CC) -I/opt/local/include -L/opt/local/lib $(CFLAGS) -o uttun uttun.c -lpthread -lssl -lcrypto
+# YMMV, but sometimes openssl etc are in /opt/local. -lssl and -lcrypto are needed only for TLS
+cbridge: cbridge.c $(OBJS) chaosd.h cbridge-chaos.h chudp.h contacts.o
+	$(CC) $(CFLAGS) $(LDFLAGS) -o cbridge cbridge.c $(OBJS) -lpthread -lssl -lcrypto
+
+contacts.o: contacts.c cbridge-chaos.h cbridge.h
+	$(CC) -c $(CFLAGS) -o $@ $<
+
+usockets.o: usockets.c cbridge.h
+	$(CC) -c $(CFLAGS) -o $@ $<
+
+chtls.o: chtls.c cbridge.h
+	$(CC) -c $(CFLAGS) -o $@ $<
+
+chudp.o: chudp.c cbridge.h chudp.h
+	$(CC) -c $(CFLAGS) -o $@ $<
+
+debug.o: debug.c cbridge.h
+	$(CC) -c $(CFLAGS) -o $@ $<
+
+chether.o: chether.c cbridge.h
+	$(CC) -c $(CFLAGS) -o $@ $<
 
 clean:
-	rm -f cbridge cbridge.o
+	rm -f cbridge $(OBJS)
+
